refactor: ルーティング方法変更
