+++ app/routes/detailList.tsx
