+++ app/routes/index.tsx
@@ -1,4 +1,4 @@
-import { Welcome } from "../welcome/welcome";
+import { Welcome } from "../components/welcome/welcome";
 
 export function meta() {
   return [
