+++ app/routes/signin.tsx
@@ -1,6 +1,7 @@
 import { commitSession, getSession } from "~/sessions.server";
 import { data, redirect } from "react-router";
-import type { Route } from "./+types/route";
+import type { Route } from "./+types/signin";
+// import type { Route } from "./signin/+types/route";
 
 export async function loader({ request }: Route.LoaderArgs) {
   const session = await getSession(request.headers.get("Cookie"));
